+++ module/item/entity.js
@@ -361,7 +361,7 @@ export class ItemPF extends Item {
 
   async update(data, options={}) {
     const srcData = mergeObject(this.data, expandObject(data), { inplace: false });
-    console.log("Updating", data)
+
     // Update name
     if (data["data.identifiedName"]) data["name"] = data["data.identifiedName"];
     else if (data["name"]) data["data.identifiedName"] = data["name"];
@@ -397,7 +397,7 @@ export class ItemPF extends Item {
         rollData.item.level = getProperty(this.data, "data.level");
         if (data["data.level"] != null && data["data.level"] !== getProperty(this.data, "data.level"))
           rollData.item.level = data["data.level"]
-        console.log('RollData', rollData)
+
         data["data.timeline.total"] = new Roll(rollFormula, rollData).roll().total;
       }
     }
@@ -425,7 +425,6 @@ export class ItemPF extends Item {
 
     const diff = diffObject(flattenObject(this.data), data);
     if (Object.keys(diff).length) {
-      console.log("Finished updating", data)
       return super.update(diff, options);
     }
 
@@ -776,6 +775,7 @@ export class ItemPF extends Item {
         primaryAttack = true,
         useMeasureTemplate = false,
           useAmmoId = "none",
+          useAmmoDamage = "",
         rollMode = null;
       // Get form data
       if (form) {
@@ -795,7 +795,10 @@ export class ItemPF extends Item {
 
         if (form.find('[name="ammunition-id"]').val() !== undefined) {
           useAmmoId = form.find('[name="ammunition-id"]').val()
-
+          useAmmoDamage = form.find('[name="ammo-dmg-formula"]').val()
+          if (useAmmoDamage !== '') {
+            damageExtraParts.push(useAmmoDamage);
+          }
         }
 
 
@@ -838,10 +841,10 @@ export class ItemPF extends Item {
       let attacks = [];
       if (this.hasAttack) {
         for (let atk of allAttacks) {
-      // Create attack object
-      let attack = new ChatAttack(this, atk.label);
-      attack.rollData = rollData;
-      await attack.addAttack({bonus: atk.bonus, extraParts: attackExtraParts, primaryAttack: primaryAttack});
+          // Create attack object
+          let attack = new ChatAttack(this, atk.label);
+          attack.rollData = rollData;
+          await attack.addAttack({bonus: atk.bonus, extraParts: attackExtraParts, primaryAttack: primaryAttack});
           if (this.hasDamage) {
             await attack.addDamage({extraParts: damageExtraParts, primaryAttack: primaryAttack, critical: false});
             if (attack.hasCritConfirm) {
@@ -1105,7 +1108,6 @@ export class ItemPF extends Item {
     if (options.primaryAttack === false) parts.push("-5");
     // Add bonus
 
-    console.log(options)
     if (options.bonus != null) {
       rollData.bonus = options.bonus;
       parts.push("@bonus");
@@ -1247,10 +1249,18 @@ export class ItemPF extends Item {
     let rolls = [];
     for (let a = 0; a < parts.length; a++) {
       const part = parts[a];
-      const roll = {
-        roll: new Roll([part.base, ...part.extra, ...extraParts].join("+"), rollData).roll(),
-        damageType: part.damageType,
-      };
+      let roll = {}
+      if (a === 0) {
+         roll = {
+          roll: new Roll([part.base, ...part.extra, ...extraParts].join("+"), rollData).roll(),
+          damageType: part.damageType,
+        };
+      } else {
+         roll = {
+          roll: new Roll([part.base, ...part.extra].join("+"), rollData).roll(),
+          damageType: part.damageType,
+        };
+      }
       rolls.push(roll);
     }
 
@@ -1446,7 +1456,7 @@ export class ItemPF extends Item {
        * -
        */
       let actions = ItemPF.parseAction(actionValue)
-      console.log(actions)
+
       for (let actionData of actions) {
         if (actionData.target === "self") {
           await actor.applyActionOnSelf(actionData, actor)
@@ -1807,4 +1817,201 @@ export class ItemPF extends Item {
       ? (getProperty(spellbook, `spells.spell${spellLevel}.value`) || 0)
       : (getProperty(this.data, "data.preparation.preparedAmount") || 0);
   }
+
+  static getMinimumCasterLevelBySpellData(itemData) {
+    const learnedAt = getProperty(itemData, "learnedAt.class").reduce((cur, o) => {
+      const classes = o[0].split("/");
+      for (let cls of classes) cur.push([cls, o[1]]);
+      return cur;
+    }, []);
+    let result = [9, 20];
+    for (let o of learnedAt) {
+      result[0] = Math.min(result[0], o[1]);
+
+      // Hardcoding classes... this seems stupid. This probably is for spell DC.
+      // We assume High
+      result[1] = Math.min(result[1], 1 + Math.max(0, (o[1] - 1)) * 2)
+      // const tc = CONFIG.PF1.classCasterType[o[0]] || "high";
+      // if (tc === "high") {
+      //   result[1] = Math.min(result[1], 1 + Math.max(0, (o[1] - 1)) * 2);
+      // }
+      // else if (tc === "med") {
+      //   result[1] = Math.min(result[1], 1 + Math.max(0, (o[1] - 1)) * 3);
+      // }
+      // else if (tc === "low") {
+      //   result[1] = Math.min(result[1], 4 + Math.max(0, (o[1] - 1)) * 3);
+      // }
+    }
+
+    return result;
+  }
+
+
+
+  static async toConsumable(origData, type) {
+    let data = duplicate(game.system.template.Item.consumable);
+    for (let t of data.templates) {
+      mergeObject(data, duplicate(game.system.template.Item.templates[t]));
+    }
+    delete data.templates;
+    data = {
+      type: "consumable",
+      name: origData.name,
+      data: data,
+    };
+
+    const slcl = this.getMinimumCasterLevelBySpellData(origData.data);
+
+    // Set consumable type
+    data.data.consumableType = type;
+
+    // Set name
+    if (type === "wand") {
+      data.name = `Wand of ${origData.name}`;
+      data.img = "systems/D35E/icons/items/magic/generated/wand-low.png";
+      data.data.price = Math.max(0.5, slcl[0]) * slcl[1] * 750;
+      data.data.hardness = 5;
+      data.data.hp.max = 5;
+      data.data.hp.value = 5;
+    }
+    else if (type === "potion") {
+      data.name = `Potion of ${origData.name}`;
+      data.img = "systems/D35E/icons/items/potions/generated/med.png";
+      data.data.price = Math.max(0.5, slcl[0]) * slcl[1] * 50;
+      data.data.hardness = 1;
+      data.data.hp.max = 1;
+      data.data.hp.value = 1;
+    }
+    else if (type === "scroll") {
+      data.name = `Scroll of ${origData.name}`;
+      data.img = "systems/D35E/icons/items/magic/generated/scroll.png";
+      data.data.price = Math.max(0.5, slcl[0]) * slcl[1] * 25;
+      data.data.hardness = 0;
+      data.data.hp.max = 1;
+      data.data.hp.value = 1;
+    }
+    else if (type === "dorje") {
+      data.name = `Dorje of ${origData.name}`;
+      data.img = "systems/D35E/icons/items/magic/generated/droje.png";
+      data.data.price = Math.max(0.5, slcl[0]) * slcl[1] * 750;
+      data.data.hardness = 5;
+      data.data.hp.max = 5;
+      data.data.hp.value = 5;
+    }
+    else if (type === "tattoo") {
+      data.name = `Tattoo of ${origData.name}`;
+      data.img = "systems/D35E/icons/items/magic/generated/tattoo.png";
+      data.data.price = Math.max(0.5, slcl[0]) * slcl[1] * 50;
+      data.data.hardness = 1;
+      data.data.hp.max = 1;
+      data.data.hp.value = 1;
+    }
+    else if (type === "powerstone") {
+      data.name = `Power Stone of ${origData.name}`;
+      data.img = "systems/D35E/icons/items/magic/generated/crystal.png";
+      data.data.price = Math.max(0.5, slcl[0]) * slcl[1] * 25;
+      data.data.hardness = 0;
+      data.data.hp.max = 1;
+      data.data.hp.value = 1;
+    }
+
+
+    // Set charges
+    if (type === "wand" || type === "dorje") {
+      data.data.uses.maxFormula = "50";
+      data.data.uses.value      = 50;
+      data.data.uses.max        = 50;
+      data.data.uses.per        = "charges";
+    }
+    else {
+      data.data.uses.per = "single";
+    }
+
+    // Set activation method
+    data.data.activation.type = "standard";
+
+    // Set measure template
+    if (type !== "potion" && type !== "tattoo") {
+      data.data.measureTemplate = getProperty(origData, "data.measureTemplate");
+    }
+
+    // Set damage formula
+    data.data.actionType = origData.data.actionType;
+    for (let d of getProperty(origData, "data.damage.parts")) {
+      d[0] = d[0].replace(/@sl/g, slcl[0]);
+      d[0] = d[0].replace(/@cl/g, slcl[1]);
+      data.data.damage.parts.push(d);
+    }
+
+    // Set saves
+    data.data.save.description = origData.data.save.description;
+    data.data.save.dc = 10 + slcl[0] + Math.floor(slcl[0] / 2);
+
+    // Copy variables
+    data.data.attackNotes = origData.data.attackNotes;
+    data.data.effectNotes = origData.data.effectNotes;
+    data.data.attackBonus = origData.data.attackBonus;
+    data.data.critConfirmBonus = origData.data.critConfirmBonus;
+
+    // Determine aura power
+    let auraPower = "faint";
+    for (let a of CONFIG.D35E.magicAuraByLevel.item) {
+      if (a.level <= slcl[1]) auraPower = a.power;
+    }
+    if (type === "potion") {
+      data.img = `systems/D35E/icons/items/potions/generated/${auraPower}.png`;
+    }
+    // Determine caster level label
+    let clLabel;
+    switch (slcl[1]) {
+      case 1:
+        clLabel = "1st";
+        break;
+      case 2:
+        clLabel = "2nd";
+        break;
+      case 3:
+        clLabel = "3rd";
+        break;
+      default:
+        clLabel = `${slcl[1]}th`;
+        break;
+    }
+    // Determine spell level label
+    let slLabel;
+    switch (slcl[0]) {
+      case 1:
+        slLabel = "1st";
+        break;
+      case 2:
+        slLabel = "2nd";
+        break;
+      case 3:
+        slLabel = "3rd";
+        break;
+      default:
+        slLabel = `${slcl[1]}th`;
+        break;
+    }
+
+    // Set description
+    data.data.description.value = await renderTemplate("systems/D35E/templates/internal/consumable-description.html", {
+      origData: origData,
+      data: data,
+      isWand: type === "wand" || type === "dorje",
+      isPotion: type === "potion" || type === "tattoo",
+      isScroll: type === "scroll" || type === "powerstone",
+      auraPower: auraPower,
+      aura: (CONFIG.D35E.spellSchools[origData.data.school] || "").toLowerCase(),
+      sl: slcl[0],
+      cl: slcl[1],
+      slLabel: slLabel,
+      clLabel: clLabel,
+      config: CONFIG.D35E,
+    });
+
+    return data;
+  }
+
+
 }